Add a parent pricing model to allow subclasses to easily get pricing info
